Reorganized the files into the corresponding folders
- Organized the files in the components  section into logical folder for easy access and lookup
