+++ frontend/newsbridge/src/components/Article/Subcomponents/BiasScale.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from "react";
-import { Button } from "./FormElements";
+import { Button } from "../../Form/FormElements";
 
 function getColor(rating: number): string {
   const red = Math.min(255, (rating / 100) * 255);
