+++ frontend/newsbridge/src/components/Article/Subcomponents/LikeButton.tsx
