+++ frontend/newsbridge/src/components/Search/SearchComponent.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useRef, useEffect } from 'react';
-import { Button } from '../FormElements';
+import { Button } from '../Form/FormElements';
 import { SortAsc, Newspaper, Clock, Gauge } from "lucide-react";
 import Dropdown from './Dropdown';
