+++ frontend/newsbridge/src/pages/component_testing_pages/BiasScalePage.tsx
@@ -1,4 +1,4 @@
-import { NewsBridgeBiasScale, UserBiasScale } from "../../components/BiasScale";
+import { NewsBridgeBiasScale, UserBiasScale } from "../../components/Article/Subcomponents/BiasScale";
 
 function BiasScalePage() {
   const handleBiasUpdate = (rating: number) => {
